+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog][homepage] and this project adheres to
 [homepage]: https://keepachangelog.com/en/1.0.0/
 [versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [1.2.1] - 2019-11-02
 
 ### Added